kvutils: Move unrelated test cases to a proper test class [KVL-1015] CHANGELOG_BEGIN CHANGELOG_END
